refactor(test,plugin): scaffold prep — externalize stubs, reserve DSL slots, init migration notes
