+++ docs/migration-v0.4-notes.md
@@ -1,0 +1,11 @@
+# v0.3 → v0.4 Migration Notes (working draft)
+
+각 feature task 가 자기 변경사항을 한 줄씩 append 한다. task 24 가 종합본 작성.
+
+## DSL changes
+
+## Generated SDK shape changes
+
+## Annotations newly recognized
+
+## Breaking changes
