+++ gradle-plugin/src/main/kotlin/io/spia/gradle/SpiaExtension.kt
@@ -7,4 +7,11 @@ interface SpiaExtension {
     val enumStyle: Property<String>
     val longType: Property<String>
     val apiClient: Property<String>
+
+    // Reserved slots for upcoming tasks — uncomment and implement when the task lands:
+    // var schemaOutput: String = "none"          // task 07 (Zod schema output path)
+    // var openApiOutput: String = "none"         // task 12 (OpenAPI spec output path)
+    // var splitByController: Boolean = false     // task 18 (split output per controller)
+    // val clientOptions: ClientOptions           // task 08-10 (nested client DSL block)
+    // val npmPackage: NpmPackageOptions?         // task 21 (npm publish options)
 }
